requirements.txt
fastapi
uvicorn
python-multipart
pillow
pillow-heif
jinja2
aiofiles
